% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GA_Route}
\alias{GA_Route}
\title{Vehicle routing using Genetic algorithm.}
\usage{
GA_Route(loc_data, dist_mat, min_cap, init, sug_sol = NULL,
  mxiter = 50000, mut_rate = 0.2, crs_rate = 0.8, converge = 5000,
  re_eval = T)
}
\arguments{
\item{loc_data}{A data frame with the nodal points and the corresponding number of people to be dropped/picked up from the nodes.}

\item{dist_mat}{The distance matrix to be used. Dimensions should be Px(P+1), where P is the number of nodes.}

\item{min_cap}{The minimum number of people that should be allotted to each vehicle.}

\item{init}{The name of the starting point/node.}

\item{sug_sol}{Optional suggested solution to initiate the GA. The vectorised output from the LS_Route function should ideally be used.}

\item{mxiter}{Maximum number of iterations for which the GA should be run. 50000 by default.}

\item{mut_rate}{The rate of mutation to be used. 0.2 by default.}

\item{crs_rate}{The rate of crossover to be used. 0.8 by default.}

\item{converge}{The maximum number of iterations to be run before stopping the GA if there is no improvement in the solution. 5000 by default.}

\item{re_eval}{Should the GA solution be re-evaluated for further optimisation.}
}
\value{
A list with the optimised vehicle routes and the total distance.
}
\description{
GA is implemented to determine the best routes so as to minimise the total distance covered by all vehicles.
The algorithm converges much faster and presents better results when the vectorised output from the LS_Route function
is provided as a suggested solution.
}
