% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LS_Route}
\alias{LS_Route}
\title{Vehicle routing using local search}
\usage{
LS_Route(loc_data, loc_code, min_cap, dist_mat, time_mat, init)
}
\arguments{
\item{loc_data}{A data frame with the nodal points and the corresponding number of people to be dropped/picked up from the nodes.}

\item{loc_code}{The column number/name in dist_mat which contains the nodal points.}

\item{min_cap}{The minimum number of people that should be allotted to each vehicle.}

\item{dist_mat}{The distance matrix to be used. Dimensions should be Px(P+1), where P is the number of nodes.}

\item{time_mat}{The time matrix to be used. Dimensions should be Px(P+1), where P is the number of nodes.}

\item{init}{The name of the starting point/node.}
}
\value{
A list containing tables for each cab allocation along with distance and time computed.
The list also contains a vectorised solution which can be fed into the GA based routing algorithm.
}
\description{
A local search algorithm is implemented to determine the optimal routes.
The result can be fed into the Genetic algorithm based routing function for further optimisation.
}
